[FIX] Fix logger tests after the refactoring
During the refactoring of the logger lib it was discovered that
test for the level ERROR  was false positive i.e. it was green but
was not verifying what it shall be, as the number of entries was
matching the expected ones.

As level ERROR component is only writing to the console, test must
have been adapted to reflect this.

What is more description of the test was fixed as well.