[FIX] Delete test_different_clients_logging
`test_different_clients_logging` was failing sporadically, due to the
consecutive calls to the `logs.get_match_in_line`.

Every time `logs.get_match_in_line` is called, it advances over the log
file, and since log clients are logging in parallel, the order of the
log entries is not fixed. Thus if one of the checks was looking for an
entry that was already read, it would fail the test.

Fixes:
SEOS-1612